requirements.txt
# Core
langgraph>=0.4.0
langchain>=0.3.0
langchain-google-genai>=2.1.0
langchain-openai>=0.3.0

# API Backend
fastapi>=0.100.0
uvicorn>=0.23.0
slowapi>=0.1.9

# Utilities
python-dotenv>=1.0.0
requests>=2.32.0
pandas>=2.2.0

# Testing
pytest>=8.0.0
httpx>=0.27.0
